update main spring 4

+++ Backend/src/main/java/com/grupo4/nos_fuimos/service/CategoriaService.java
@@ -28,29 +28,43 @@ public class CategoriaService {
         return categoriaRepository.findAll();
     }
 
-    public Categoria addCategoria(Categoria categoria){
-        return categoriaRepository.save(categoria);
+    public ResponseEntity<?> addCategoria(Categoria categoria){
+
+        Optional<Categoria> existingCategoria = categoriaRepository.findByTitulo(categoria.getTitulo());
+        if (existingCategoria.isPresent())
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("La categoria con nombre '" + categoria.getTitulo() + "' ya existe en la base de datos");
+
+        else{
+            Categoria savedCategoria = categoriaRepository.save(categoria);
+            return ResponseEntity.status(HttpStatus.CREATED).body(savedCategoria);
+        }
     }
 
     public Categoria getCategoriaById(String id){
         return categoriaRepository.findById(id).orElse(null);
     }
 
-    public Categoria updateCategoria(List<String> productosId, String idCategoria){
-        Categoria categoriaEncontrada = getCategoriaById(idCategoria);
-        if (categoriaEncontrada != null){
-            categoriaEncontrada.setIdProductos(productosId);
-            return categoriaRepository.save(categoriaEncontrada);
+
+    public ResponseEntity actualizarCategoria(Categoria categoria) {
+        Optional<Categoria> optionalCategoria = categoriaRepository.findById(categoria.getId());
+        if(optionalCategoria.isPresent()){
+            return ResponseEntity.ok(categoriaRepository.save(categoria));
         }
-        return null;
+        else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Categoria no encontrada");
+        }
     }
 
-    public void deleteCategoria(String id){
-        categoriaRepository.deleteById(id);
+    public ResponseEntity eliminarCategoriaById(String id){
+        Optional<Categoria> optionalCategoria = categoriaRepository.findById(id);
+        if(optionalCategoria.isPresent()) {
+            categoriaRepository.deleteById(id);
+            return ResponseEntity.ok("Categoria eliminada correctamente");
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Categoria no encontrada");
+        }
     }
-
-
-    }
+}
